Server Upload(Server
Server end of the server upload. Once socket implementation is ready on the client end I will convert this slightly to provide the game page via the server. This will make the socket connection inherent in the actual communication. Validation code is almost ready, those versions of the code will be up ASAP