First stage for server Update
A few very large changes. I have reorganized the structure of the code so that the server itself is providing the game page to the user. Need to make it so that everything is feeding the right bits to the right places but other than that this is "ready" to be set live at any time. Collision still needs to be implemented, and validation was put on the back burner to figure this out.
FOR TESTING: So right now to implement a lot of the socket shit, we need two computers to be using node.js servers configured for the socket (Big ask), or a coordinated time between myself and Mubeen/Albeer, Client side will be edited on the fly and sent to me while I implement server end implementations of the code. Even with the http tunnel it won't make editing it make sense since I can only open my server to everyone, can't open my IDE, well without giving up control of my laptop. Need to discuss this in the meeting today.