Stage 2 of server update commit
NOTE: A lot of this code won't work unless someone attempting to test this is using .env. package.json lists all of the needed packages for the user to install. These are .js bits that are going to be doing grunt work for the app.js file, otherwise it is too cluttered. Honestly, might try and farm out the socket code into another file just to shorten it up.