Validation Complete!
Finally we have fully functioning validation my friends. User registration is fully integrated with blowfish hash functions to make it difficult for intruders to access passwords even if we get all the data taken. Account verification is based on the username however, email notification is eventually going to come if possible. Leaderboard is coming along, worst case scenario I dump the rest of that code that isn't working in here anyways. But other than that, all that needs to happen is the http tunnel gets activated and then pointed to the port I have set and it is live.