Leaderboard done ajax calls complete
Leaderboard is implemented at long last. 
The tragedy of socket being verbotten was not overcome today. However we are in the process of getting a very robust system to allow for multiplayer experience with the resources at out disposal.
More in depth gameplay, the user will be getting a life bar and the user will be receiving dangerous balls thrown at them. The player can day and there is a constantly running leaderboard to see the current top of the game.