Leaderboard and all that
See the commit two commits ago for details, leaderboard is ordered fully and prints to the screen nicely, the pages redirect pretty nicely and smoothly, ajax queries are quick and while tested on the http tunnel quite reliable. Emergency updates tomorrow and this is ready to ship, albeit missing a crucial feature but such is life.